backend/collector/requirements.txt
redis
asyncio
playwright
minio
psycopg[binary]
feedparser
Mastodon.py
PyGithub
requests
